SPEAr : Network driver fixes.
The network controller mac resets hardware address stored in MAC_HI and MAC_LO
registers if mac is resetted. So, hw mac address needs to be restored in case
mac is explicitly resetted from driver.
Also the max frame length for Normal descriptor can be 0x7FF i.e 2047, so
reducing it to 1600

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>