[SH] Bug-fix, to display cut number correctly for Freeman SoCs
This update will calculate the major cut number for the FLI7510 and
FLI7540 SoCs from STMicroelectronics. Before this update, the cut numbers
would be deemed to be one higher than they should. That is a cut 0.x
part would be deemed to be a cut 1.x part. This update fixes this error.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>