[SH] Added support for boot-from-SPI on the FUDB (FLI7540)
This update adds support for boot-from-SPI on the Freeman Ultra
Development Board (FUDB), which uses STMicroelectronics' FLI7540 SoC.

In addition, support for NAND "bit-banging" has been removed, as a review
of the schematics has shown that it is not possible to use the NAND in
bit-banging mode. This is because the ALE and CLE signals are not brought
out on the EMI address lines, but are controlled directly by the (FLEX)
NAND controller.

Note: this was tested on a single FUDB board ("FUDB_GPD201"), with serial
number SHZ019, which had a board modification, to work around a known
problem with setting the "mode pins" for boot-from-SPI.

The remaining outstanding features are: SPI serial flash with the FSM
driver, and to get the "reset" command to work properly.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>