[SH] Added the file 'include/asm-sh/sysconf.h'
This update adds a new header file 'include/asm-sh/sysconf.h', which
helps when setting then system configuration registers. It is simply just
a pair of macros to set/clear a single or a set of contiguous bits in
a variable.

	/*
	 * if ('flag')
	 *	set bit 'bit' in variable 'reg'
	 * else
	 *	clear bit 'bit' in variable 'reg'
	 */

	/*
	 * if ('flag')
	 *	set bits 'lsb:msb' to 'yes' in variable 'reg'
	 * else
	 *	set bits 'lsb:msb' to 'no' in variable 'reg'
	 *
	 * Note: 'msb' must be >= 'lsb'.
	 */

Note: this commit should logically change nothing, it is just a minor
refactoring operation, to move some duplicate definitions into a single
place.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>