[SH] Add Ethernet support for the STx7108 (MB837)
This update adds support for the two (on-chip) STMACs, which may be
attached to a suitable daughter-board containing a PHY.

Both STMACs were tested (in turn), when attached to a MB859A board,
which contains an IC+ IP1001 PHY. Both MAC+PHY combinations were found
to work, with no problems.

Please note the following:

The ethernet PHYs may be reset by one of two mechanisms:

	a) the EPLD on the MB705 peripheral board

	b) via the PCF8575 (IC12) PIO Extender.

This code does *not* (currently) make any attempt to reset the PHYs,
but relies on the MB705 to reset the PHYs on power-on, which (thus far)
demonstrably works well!

If an MB705 is *not* connected, then additional work is required,
including fitting J14-A, and J14-B.  We currently just issue a diagnostic
to warn about this ...

Testing was performed on a MB837A (cpu board) mated with a MB705B
(peripheral board), together with a pair of MB859A (IC+ PHY)
daughter-boards.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>