[SH] Added support for the HDK-7108 board (STx7108).
This update adds initial support for STMicroelectronics' HDK7108 board,
which uses STMicroelectronics STx7108 SoC.

This update includes support for the following:

	32-bit (SE) mode, RS232 serial, NOR Flash, NAND Flash, I2C,
	Ethernet (both GMACs), boot-from-NOR, and boot-from-NAND.

The following are not yet supported: SPI, USB, SATA.

All testing was done on a single board: STi7108-HDK10ST No50.

Note, it should be noted that on power-on, the board does *not* boot.
However, booting does succeed if the "reset"  button (near the RS-232)
is pressed. This may be due to a H/W bug with the board's power-on
reset circuitry.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>