[SH] Use the "UPDATE32" opcode (instead of POKE32) to set the SDRA registers
This update is to address a problem on some of the older boards that
STMicroelectronics supports, whereby when in 32-bit (SE) mode, the LMI's
SDRA register value could replace *all* the bits, and not *some* of the
bits when performing the standard set of pokes.

Using the "UPDATE32" opcode instead of "POKE32" will preserve the
bottom 21 bits, when updating the UBA bitfield (located in the top 11
bits). The failure situation we are trying to address is were GDB's
targetpacks are using the latest values, and when romgen is run, the new
".romgen" file will use these latest values (as it should).  However, in
32-bit mode, U-Boot (currently) needs to "over-write" the UBA bitfield,
but should leave the other bitfields alone. On the more recent boards,
this semantic is honoured.  However on some of the older boards, a naive
"POKE32" was used.  This update rectifies this situation, and U-Boot
will now perform an "UPDATE32" on all relevant boards.

Note:	No testing was performed for this commit, other than building
	U-Boot for the affected boards.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>