[SH] Get the "reset" command to work for boards using the STx7108 SoC.
On the STx7108 the default value of the reset pulse length is longer
than the WDT time-out - so we end up re-triggering the WDT before the
chip comes out of reset the 1st time - resulting in an endless loop!
So, we make sure reset period is shorter than WDT time-out. We now set
a time-out of about 100us.

With this modification, the "reset" command now correctly reboots U-Boot.
This was tested on a single MB837A+MB705B, with boot-from-NOR flash.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>