[SH] Added initial support for the MB903 board (STx7108).
This update adds initial support for STMicroelectronics' MB903 board,
which uses STMicroelectronics STx7108 SoC.

This update includes support for the following:

	32-bit (SE) mode, RS232 serial, NOR Flash, I2C, Ethernet (GMAC #1)

The following are not yet supported: SPI, USB, SATA.

All testing was done on a single board: MB903A2 N0001

Note, the NOR testing is "incomplete", as the board seems "unreliable"
as far as NOR accesses is concerned. Hence, boot-from-NOR has not been
attempted yet.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>