[SH] Added SPI and boot-from-flash support for the MB903 board
This update adds SPI (bit-banging) support for the MB903, which uses
STMicroelectronics' STx7108 SoC. (It is not possible to use the SSC.)

In addition, "romgen" was run, and the appropriate poke tables is now
also included. As a result, both boot-from-SPI and boot-from-NOR have
been successfully tested.

All testing was performed on a single board MB903A2 N0001.

It should be noted that when in boot-from-SPI mode, that only the upper
64 MiB of NOR flash is accessible.

Only remaining significant tasks are SATA & USB (OHCI).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>