[SH] Perform a double-fault to effect a "reset" on SH4-300 series cores.
Modify SH's "reset" command, so that it will now perform a double-fault
(i.e. "TRAPA" instruction, with SR.BL==1) on SH4-300 series cores.

A double-fault is more desirable than a Watchdog time-out, as there
is no requirement to "pacify" the Watchdog so soon after the reset
pulse is de-asserted.  On boot-from-SPI, and boot-from-NAND, where the
time taken to read the first page can take a (relatively) long time,
(especially if you need to scan 512 NAND blocks to find logical block
zero!), a double-fault should result in a more "reliable" reset.

However, for SH4-200 cores, we will continue to effect a reset via the
Watchdog timer, since the double-fault is not guaranteed to reset the
core back into 29-bit mode, if we were in 32-bit (SE) mode.

This change has only been tested on a single MB903 (STx7108) board,
booting from NAND.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>