[SH] Add support for recognising "Boot-Mode+B" ECC in NAND flash.
This update will allow NAND flash blocks which were created with the
special "Boot-Mode+B" ECC layout to be recognised still as a "good"
block, despite the fact that the "bad block" marker has been corrupted,
even without the use of Bad Block Tables (BBTs).

That is, it is possible to write to a clean NAND device (without BBTs)
data using the "Boot-Mode+B" ECC, and subsequent invocations will deem the
block still to be good, even though the bad block markers are not 0xFF.

The "Boot-Mode+B" ECC scheme uses 100% of the OOB. This ECC is also
described as "3+1/128". That is, for every 128 bytes of in-band data,
there are 3 bytes of ECC followed by 1 tag byte ('B' == 0x42) stored in
the OOB area.

Without this change, and without using BBTs, then the standard "bad
block markers" will *always* clash with the H/W Boot-mode ECC as used
by STMicroelectronics. That is, a "good" block, when written to with
Boot-mode ECC (either 3/128, or 3+1/128) will (most likely) subsequently
appear as bad, as the bad block markers will have been over-written with
ECC data! Prior to this update, the use of BBTs was the only solution.

By default, when CFG_NAND_ECC_HW3_128 is enabled, then the "Boot-Mode+B"
ECC is also now implicitly enabled. To revert, and use the pure H/W
ECC 3/128 ECC (leaving every fourth byte free), then add the following
to config.h:

	#define CFG_STM_NAND_BOOT_MODE_ECC_WITH_B 0

Note: to test this, then ensure that NAND_USE_FLASH_BBT is not set
in nand->options in board_nand_init() in nand.c in the board specific
tree. The resultant U-Boot will not use BBTs, and recognise a block as
good if either the bad block marker is 0xFF, or every 4th byte in the
OOB is ASCII 'B'.

Note: if NAND has been written to using an older version of this code
(or pure 3/128) then the additional 'B' tags will be missing, and the
blocks will not be recognised as good. The NAND device should be erased
using a contemporaneous version of U-Boot.

Note: a single bit error of all the 'B' tag bytes is tolerated, and the
block will still be recognised as good, as long as there is at most just
one bit error across all the 'B' bytes.

This allows full interoperability between STLinux and U-Boot, regarding
the use of the "Boot-Mode+B" ECC in NAND flash devices, by default.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>