[SH] Assistance with disabling non-volatile (NAND-resident) BBTs
This update is a trivial change, which should not affect behaviour at all.

All it does is make it easier in the board/st/*/nand.c files to allow
users to enable (or disable) a NAND-resident (non-volatile) Bad Block
Table (BBT).

The comments should now make it clear that the user should still override
scan_bbt(), even if they are not using a Bad Block Table (BBT). This
is required so the special NAND ECC 3+1/128 (i.e. "Boot-Mode+B") layout
(see previous commit) is still recognised, even if a volatile (RAM-based)
BBT is used, so that "good" blocks with Boot-Mode ECC are still treated
as good (despite the fact that the bad block markers may not be 0xFF).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>