[SH,BUGFIX] STM Boot-Mode+B ECC now recognised for small-page NANDs
This is a bug-fix to the penultimate commit
(0b2c5b9d47f91ab0d7fb21355874edaec7130d49) which added support for
recognising STMicroelectronics "Boot-Mode+B" (3+1/128) ECC scheme.

That commit was only tested with a large-page NAND device. Subsequent
testing with a different board using a small-page NAND device demonstrated
there was a bug which meant that the magic "B" sequence was comparing
the wrong bytes in the OOB.

This commit rectifies that bug, and has been tested using both small-page
and large-page NAND devices.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>