[SH] Add support for recognising "AFM4" ECC in NAND flash.
This update will allow NAND flash blocks which were created with the
special "AFM4" ECC layout to be recognised still as a "good" block,
despite the fact that the "bad block" marker has been corrupted, even
without the use of Bad Block Tables (BBTs).

That is, it is possible (in linux) to write to a clean NAND device
(without BBTs) data using the "AFM4" ECC, and subsequent invocations
will deem the block still to be good, even though the bad block markers
are not 0xFF.

The "AFM4" ECC scheme uses 7/16 of the OOB. This ECC is also described
as "4+3/512". That is, for every 512 bytes of in-band data, there are
4 bytes of ECC with 3 tag bytes (ASCII "AFM") stored in the OOB area.

Without this change, and without using BBTs, then the standard
"bad block markers" will always clash with the AFM4 ECC as used by
STMicroelectronics. That is, a "good" block, when written to with AFM4
ECC will (most likely) subsequently appear as bad, as the bad block
markers will have been over-written with ECC data! Prior to this update,
the use of BBTs was the only solution.

By default, any blocks with the AFM4 (4+3/512) ECC pattern will now be
recognised as being a "good" block when NAND is scanned for bad blocks.
To defeat this new capability, then add the following to config.h:

	#define CFG_STM_NAND_AFM4_ECC_WITH_AFM 0

Note: to test this, then ensure that NAND_USE_FLASH_BBT is not set
in nand->options in board_nand_init() in nand.c in the board specific
tree. The resultant U-Boot will not use BBTs, and recognise a block as
good if either the bad block marker is 0xFF, or the "AFM tag is seen at
offset 0x3 in every 16-byte record in the OOB.

Note: a single bit error of all the "AFM" tag bytes is tolerated, and
the block will still be recognised as good, as long as there is at most
just one bit error across all the "AFM" bytes.

Please see the antepenultimate commit (0b2c5b9d47f91ab0d7fb21355874edaec7130d49).

It should be noted that (as of the time of writing) that U-Boot does *not*
support either AFM mode, nor the AFM4 ECC. That is, although U-Boot will
now recognise blocks with AFM4 ECC as being "good", U-Boot is unable to
read or write such blocks. However, such blocks written by STLinux will
be recognised as such.

An update to U-Boot to support native reading/writing of AFM4 is planned.

This was tested on a PDK-7105 board, on which linux had written blocks
with both "Boot-Mode+B" ECC and "AFM4" ECC withOUT a BBT.  U-Boot could
successfully recognise those blocks, and build a valid BBT.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>