[FAT,BUGFIX] Fixes a bug in the FAT32/FAT12 filesystem (cluster ranges).
The code (quite sensibly) checked that cluster numbers are "in range"
before trying to read data from them. However, the permitted range of
valid clusters was fixed at 0x0002 to 0xFFEF.  This range is correct
for FAT16, but *not* for FAT32 (nor FAT12).

This update adds a new function invalid_cluster(), which will test for
invalid ranges, but the range is dynamic, based on the filesystem being
used, i.e. if is is FAT12, FAT26 or FAT32.

This was tested on a FAT32 filesystem with 100k+ clusters, and "fatload"
was used to read a file known to be outwith the valid range for FAT16
clusters.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>