[NAND] Force "nand erase" arguments to be block aligned.
This update will issue additional diagnostics if, either the start
address or the requested length, is not aligned on a NAND erase block
size, when running "nand erase". If such a diagnostic is generated,
then the command will fail, without any blocks being erased.

This should avoid the case where a user perhaps only expects (say)
a single page to be erased, but U-Boot (prior to this update) would
(silently) "round up", and erase the entire block! By forcing the user
to align explicitly on NAND block boundaries, this should make the user
more aware of NAND's idiosyncrasies.

Please see:
	https://bugzilla.stlinux.com/show_bug.cgi?id=9759

This now makes the "nand erase" command more consistent with the "nand
write" command - which forces *page* alignment for writing to NAND.

To be clear:
	"nand write" arguments (start+length) must be page aligned
	"nand erase" arguments (start+length) must be block aligned

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>