[SH] Rename phy_reset() to stmac_phy_reset() and define it "extern"
Minor change, with no modification in behaviour.

For the ST40 files, just rename the function phy_reset() to
stmac_phy_reset(), and make it "extern", so it may be used in other
functions. New function prototype now in include/asm-sh/soc.h.

The rationale for this, is to help with measuring "boot-times".  That is,
we can now put a probe on the ETH_RESET signal, and measure the time it
takes from CPU_RESET to the ETH_RESET transition, and this is the elapsed
"boot-time".

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>