[SH] Add signalling mechanism to measure the time to boot linux.
In order to measure boot-times reasonably accurately, it is desirable to
toggle some pin/signal to indicate that the period we wish to measure has
started and terminated. We can then attach some probes from a digital
oscilloscope (or logic analyser) and measure the time between the two
transitions.

In the case of measuring the time for U-Boot from "reset" to "ready
to pass control to linux", then just before we pass control to linux,
would be a good opportunity to toggle a PIO pin to indicate that we can
stop measuring time. We ideally want to do this before we irretrievably
"damage" the PMB configuration, if we are in 32-bit (SE) mode.

In this case, we call stmac_phy_reset() before retuning to the caller. It
should be noted that although we are calling stmac_phy_reset(),
our intention is *not* to reset the PHY, but it is purely to effect a
transition on a pin that we can easily add a probe to. The fact that the
PHY has been reset is just be a benign by-product of this "signalling"
mechanism.

This mechanism is disabled by default, in order to enable this mechanism,
then the "#if 0" needs to be modified to include the new code.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>