Additional diagnostic when we have probed an unrecognised NAND device
This commit just prints out the Manufacturer's ID and the Device ID when
an "unknown" NAND device is probed. This is useful when we are starting
to get boards with MLC NAND devices, but (current) U-Boot does not know
what they are.

This commit can be removed once proper support for MLC NANDs has been
incorporated.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>