[SH] Add support for STx7108 cut 2.x (boot-from-NOR) on the HDK7108
This update adds support for cut 2.x of STMicroelectronics' STx7108 SoC,
on STMicroelectronics' HDK7108 (Revision 2.x) boards. The same binary
image should work on all boards with either cut 1.x or cut 2.x parts.

This was tested on several HDK7108 boards as follows:

	Cut	Board Revision	Board Serial
	----	--------------	------------
	1.1	V1.0		N176
	2.0	V2.0		N378, N385, N401
	2.0	V2.1A		N412

Note: With the V2.0 boards, the following TargetPack options may be
required when using V10 of the TargetPack:

		tck_init_frequency=1000000
		boardrev=1

Only boot-from-NOR was tested, as U-Boot currently does not support the
MLC NANDs which are present in the newer Rev 2 boards. Support for MLC
NAND will be provided later.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>