[SH] Bug-fix erroneous "time-outs" when writing FLASH.
This commit fixes a bug in lib_sh/time.c, whereby the function get_timer()
could return a large number (with msb set), due to the H/W counter
over-flowing. This happened with driver code like the following:

	unsigned long base = get_timer(0);
	function(); /* do something */
	unsigned long delta = get_timer(base);

If the first call to get_timer() was just before the H/W TMU counter
over-flowed (wrapped around) and the second call was just after, then we
would effectively calculate the displacement as a negative time interval,
which resulted in subsequent code erroneously believing a time-out had
in fact occurred, and aborting the current operation.

This problem occasionally manifested itself with spurious "timeouts"
when erasing/writing FLASH devices, or performing NFS transfers.

This was tested on a single STi7108-HDK (Rev 1.0) board, by first erasing
and then successfully programming 63MiB of zeros to a NOR Flash device.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>