[SH] bug-fix to tick_delay() in lib_sh/time.c (stopping a hang)
The function udelay() calls tick_delay() to wait for the specified number
of H/W clock ticks.

The code in tick_delay() was careful to ensure that if the computed 'end'
time was greater than the maximum H/W counter, that it would first wait
for the counter to over-flow, and then re-calculate the new 'end' time
(by masking of the top 32-bits) - fine.

Unfortunately, there was a flaw in the code, which (very) rarely could
result in tick_delay() effectively looping for ever! The problem was if
the computed 'end' time was equal to (or just under) the maximum counter
value, then it could be almost impossible (empirically) to satisfy this
condition, and the function would effectively loop forever. This failure
was observed with 'end' times of 0xfffffffe and 0xffffffff.

This commit fixes this problem by terminating in the faulty case, either
if the 'end' time is exceeded, OR the 'end' time is less than the 'start'
time - the latter clearly being indicative of an overflow having occurred.

This was tested by calling "udelay(1)" in a tight-loop, and letting this
run for 4 days, with no observable 'hang' being detected.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>