[SH] Bug-fix to improve SATA HDD Detection on STM's SoCs.
This commit addresses the issue that SATA HDDs are sometimes not
detected (Seagate ST3500312CS or WD2500AVVS) on the STx7141.  Please see
https://bugzilla.stlinux.com/show_bug.cgi?id=10162 for full details.

As I do not have access to a environment to reproduce the problem, then
my testing was just to ensure that the code builds, and that on one board
(PDK-7105) that there were no regressions with SATA. U-Boot continued
to work fine irrespective if a SATA HDD was present or not.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>