[SH] Minor enhancement to previous commit (perform dummy peek)
This update is a minor enhancement to the previous commit. Please read
the previous commit for the appropriate context.

This update will perform the "dummy peek" to the SPI serial flash after
writing to the SPI_MODESELECT register, during the initialization in
start.S, if we are booting from SPI serial flash.

This update also removes the need for the TargetPacks to perform the
"dummy peek" via JTAG in spi_configure().

This was tested on a single V2.0 board (STi7108-HDK20ST-N359), with a cut
2.0 of the STx7108 SoC present. Boot-from-SPI was successfully tested,
booting from both Flash and JTAG

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>