[SH] bug-fix to "bit-banging" SPI on MB903
This commit fixes a bug that was introduced in the penultimate commit.
That commit added support for the SPI FSM for the STx7108 (on the HDK-7108
board), but accidentally broken the ability to use SPI "bit-banging"
on the MB903.

This commit now allows SPI bit-banging on the MB903 to work (as it was)
prior to the penultimate commit.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>