[SH] adds (putative) support for FSM SPI on the STx7108M cut 2.x (MB903)
This update adds (putative) support for the FSM SPI controller on
STMicroelectronics' STx7108M SoC for cut 2.x, on the MB903A board.

It should be noted, that at the time of testing, there was no local
access to a STx7108Mc2 SoC, so this update has been made blind, and as
such is untested for that configuration. As a result, the default SPI
mode is still set to be S/W "bit-banging", and not H/W "FSM", which
should (logically) work in cut 2.0 devices.

However, in the process of implementing this new change, an extant bug
was observed with the S/W "bit-banging" implementation when U-Boot's
environment was stored in SPI serial flash. This has now been rectified.

SPI serial flash (with "bit-banging") using the STx7108M (cut 1.1)
has now been successfully tested on a single board (MB903A2-N006).
This was tested with both JTAG downloads, and real boot-from-SPI.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>