[bug-fix] Remove "inline" from "weak" function definitions
It does not really make sense to inline a weak function, since this
would prevent it to be preempted at link time.

This patch simply removes the "inline" attribute from such functions.

This was required, as the new version of GCC (4.4.4) complains with an
error, if such a function is encountered.

	See https://bugzilla.stlinux.com/show_bug.cgi?id=10476

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>