[GCC] Fix strict aliasing issues for compatability with GCC 4.4.4
The new version of GCC (version 4.4.4) is more aggressive in the way
that it treats "strict aliasing issues" (e.g. -Wstrict-aliasing).

As a result, when U-Boot is compiled with GCC 4.4.4, there are several
new diagnostics emitted pertaining to strict aliasing issues, which were
not emitted with the previous version of GCC (4.2.4).

This update fixes these issues, by either modifying the code to
comply with the strict(er) aliasing rules, or (in one case) pass the
"-fno-strict-aliasing" option to the compiler.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>