[SH] Added support for HoM (Hibernation on Memory) for HDK-7108
This update adds initial support for Linux's HoM (Hibernation on Memory)
capability, which may be used to implement "Controller Passive Standby".

This support is initially only been tested on STMicroelectronics' STx7108
(cut 2.0) SoC device, on the HDK-7108 board.

By default, HoM support is disabled. Hence in order to enable it, the
following macro needs to be defined:

	/*
	 * Define the following only if we want to use the
	 * Linux Hibernation-on-Memory (HoM) capability.
	 */

In addition, both U-Boot and Linux need to share a common "tag" in RAM,
so that U-Boot knows to restart linux via the HoM mechanism. To define
the address of that tag, the following macro needs to be defined:

	/*
	 * If we are using HoM, we need to tell U-Boot where the
	 * special "tag" is located in memory. This must be the
	 * same address as configured for the linux kernel.
	 */

Note: These macro names in U-Boot are the same as used by linux.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>