[SH] Write out the PHY ID if STMAC failed to find a known PHY
This is a minor enhancement for the STMicroelectronics' STMAC ethernet
driver. This update will produce additional diagnostic output if the
MAC can not find a PHY that it recognises. This is trying to address the
problem where the PHY on the board is different to the one that U-Boot
is configured with! The PHY IDs of all unrecognised PHYs (which respond)
will now be displayed. This should assist in detecting unconfigured or
new PHY devices which may be connected to the on-chip MAC.

In addition, a minor bug was also fixed, whereby the higher network
levels should no longer try and talk to the MAC if no known PHY was found.
This previously could result in the system hanging.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>