[SH] bug-fix to FLEX driver, to allow OOB to be written properly
This is a bug-fix to the FLEX NAND driver, as used on STMicroelectronics
SoC devices.

The FLEX controller requires data to be written on multiples of
4-bytes. For normal operations such as data read/write, the driver
carefully ensure that this constraint is always met.

However, a bug has recently been disinterred that shows that when the
"nand markbad" command is used (without a NAND resident BBT), that the
code will try to write the 2-byte "bad block marker" directly in the
OOB area, and the size of this write is only 2-bytes, which is results
in the bad block marker *not* being written correctly to the OOB.

This commit fixes this bug, by ensuring that the OOB data is written
into a temporary buffer first, and then written as a whole to the FLEX
controller.

This has been tested on a single PDK-7105 board (with large-page NAND).
Although untested, it should also work with small-page NAND devices.

For further details, please see:
	https://bugzilla.stlinux.com/show_bug.cgi?id=10848

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>