[NAND] Added initial support for multi-bit on-die ECC on SLC NAND devices
This update adds support specifically for the large-page Micron
MT29F2G08ABAEA SLC NAND, which uses on-die (i.e. on the NAND die) ECC
generation/corrections. This does not use the BCH engine on the SoC.

This commit has been tested only on a single (STx7141-based customer)
board which was fitted with a suitable NAND device.

It should be noted that there is *no* support for boot-from-NAND. The
board was booted via GDB, and then the on-die ECC on the NAND was then
enabled as a "feature". Also, there is no support for MLC NAND, only
SLC at this point.

This has only been tested on the aforementioned (single) large-page NAND
device. However *some* of the code (e.g. the "SET FEATURES" commands)
has also been written for use on small-page NAND devices (although it
is completely untested).

WARNING: It is known that for small-page NAND, the NAND-resident Bad
Block Tables (BBT) will not work correctly, as the "bbt" signature has
(currently) been moved to beyond the end of the OOB area! I plan to add a
specific patch to the code shortly to either issue an appropriate warning,
or fix this issue properly. However, the code (as is) is "safe" as the
on-die ECC is currently only enabled for the aforementioned NAND device,
which is large-page anyway, and hence does not suffer from this issue.

Thanks to Jolly LI <jolly.li@st.com>, for the initial development,
leading up to this commit.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>