[NAND] Add support for "nand (read|write).oob" sub-commands
This update adds support for the following 2 new "nand" sub-command:

	nand read.oob  addr off size
	nand write.oob addr off size

This is a generally useful feature to add, which was in the "legacy"
NAND support anyway. However, it was added to help with the on-going
development of the multi-bit on-die ECC on SLC NANDs, whose initial
version was just added a few commits ago.

This commit has been tested only on a single (STx7141-based customer)
board which was fitted with a suitable large-page NAND device.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>