[NAND] Completed support for MT29F2G08ABAEA (multi-bit on-die ECC SLC NAND)
This update should complete the support for Micron's MT29F2G08ABAEA SLC
NAND, which uses on-die (i.e. on the NAND die) ECC generation/corrections.

The update is to make the code explicitly aware that the OOB area has a
different layout from "normal" (or traditional) NAND devices. Specifically
the ECC bytes are stored in positions fixed by the H/W ECC engine on
the NAND device.  As a result of this, the location of the "free OOB"
bytes used by the auto-place algorithm needs to be modified. This is
especially true for things like the JFFS2 "clean block markers", which
*clashed* with the default (prior to this commit) placement locations.

That is, prior to this commit, using U-Boot's "nand erase clean"
sub-command would place some of the JFFS2 clean block marker in the same
location as used by the H/W ECC engine, and hence JFFS2 would (probably)
not work correctly.

For Micron MT29F2G08ABAEA, the 64-byte OOB is divided into 4 identical
records.  Each 16-byte record has the following layout:
	0x00 - 0x01 : Reserved (for Bad Block Markers)
	0x02 - 0x03 : User Metadata II (unprotected)
	0x04 - 0x07 : User Metadata I  (protected)
	0x08 - 0x0f : ECC for main + Metadata I regions

This commit will *only* use the UN-protected bytes for Metadata,
as modifying these bytes does *not* affect the ECC bytes, which may
require the entire block to be erased if there are any 0->1 transitions.
This means there are 4*4 (16) bytes of OOB that are "unusable".
Clarification with Micron to confirm this understanding will be sought!

This means that the JFFS2 signature will be stored in the following
eight bytes in the OOB: 0x02, 0x03, 0x12, 0x13, 0x22, 0x23, 0x32, 0x33.

Also as a result of this decision, it is understood that there are
probably insufficient spare bytes in the OOB to support YAFFS at the
moment. Should YAFFS support be needed, then we will need to analyse
the usage of the ECC protected bytes in the OOB within the context of
the YAFFS filesystem.

It should still be noted that there is *no* support for boot-from-NAND.
The board was booted via GDB, and then the on-die ECC on the NAND was
then enabled as a "feature". Also, there is no support for MLC NAND,
only SLC at this point.

On checking with Micron, it currently appears that there are no small-page
NAND devices in existence (nor should there be) which require multi-bit
ECC support. A few commits ago, I pointed out (in the commit log) that
the BBT would not work correctly, as the "bbt" signature has been moved
to beyond the end of the OOB area for small-page devices. Given the fact
that this now seems a non-problem, I have elected not to modify the code
to detect this errant condition.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>