[SH] Disable SPI (by default) to avoid "hang" on HDK-7108 boards
As reported in a previous commit 2af31e5c52142d966fc60393e02609d1a4c358fd,
is is possible for the SPI initialization driver to "hang" on
STMicroelectronics' HDK-7108 boards.

To avoid customers (in the short term) with this annoyance, this commit
simply disables (by default) support for SPI. Should users wish to use
SPI on this board, then they should define CONFIG_SPI_FLASH, and review
the commit comments for 2af31e5c52142d966fc60393e02609d1a4c358fd.

In addition, a better description of which versions of the board and
chip are supported is also included in "include/configs/hdk7108.h".
In summary, only STx7108 cut 2.0 (or later) and HDK-7108 board revision
2.0 (or later) are supported on U-Boot.

For further details, please see:
	https://bugzilla.stlinux.com/show_bug.cgi?id=11091

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>