[SH] Add support for SPI Quad-mode booting on the HDK-7108
This update adds support for the ability to boot in "Quad-mode" (i.e.
x4) using the FSM SPI serial flash controller on STMicroelectronics'
STx7108 SoC for cut 2.x, on the HDK-7108 board.

It is important to note that by default SPI quad-mode is *not* included
in this update. This update essentially redefines CFG_EMI_SPI_BASE so that
if U-Boot's environment is stored in SPI, and we are subsequently booting
from SPI, that the environment will still be found correctly. A suspected
bug may exist which meant that the environment was "missing", which can be
easily resolved using a cached mapping (instead of a un-cached mapping).
Using caching is better anyway, so this change is benign.

Please see the additional notes in "board/st/hdk7108/init-hdk7108.S" for
details on how to re-run romgen to enable quad mode. In addition, to boot
from SPI, you also need to define: CONFIG_SPI_FLASH, CFG_BOOT_FROM_SPI,
and possibly CFG_ENV_IS_IN_EEPROM.

It should be noted that this update does *not* modify U-Boot's SPI driver,
and as such there is still no support for either dual or quad mode
in the SPI driver. This lack of support in the driver need not impact
boot-time, as the faster boot-time that quad offers is achieved by the
"pokes" from romgen, and not the SPI driver.

Testing was performed on a single board (STi7108-HDK20ZB N482),
booting successfully from SPI in both x1 and x4 modes. In all cases the
environment was located in the SPI, and was found correctly, and was
written using the (x1) SPI driver.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>