[SH] bugfix for broken support for the TERIDIAN 78Q2123 PHY
This update simply fixes a bug from a recent commit
(6c11de76fada1d05369f644c5ed19c8131aefbf4) when support for the Realtek
Ethernet RTL8201E PHY was added. That commit accidentally broke support
for the TERIDIAN 78Q2123 PHY (as used on the HDK-7111).

Please note, the committer (Sean McGoogan) has not been able to test this,
modulo that it builds cleanly.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>