[SH] minor change to usage of the SH4_SET_PMB assembler macro
This update does not (or should not) change the semantics of the code
at all, and hence should be no change in logic of the code.

Whilst doing some recent changes, it was noted that in the cpu/sh/start.S
file, that there are many instantiations of the gas macro "SH4_SET_PMB".
However, strictly speaking the parameters to a macro should be
COMMA-delimited, and not SPACE-delimited. In *most* cases this makes no
difference. However, if parenthesis are to be used to force precedence
evaluation, then SPACE-delimited parameters result in assembler warnings!

That is, the following is illegal:

	SH4_SET_PMB 1 0x88 (BASE+0x08) 128 0 0 1

But the following is correct:

	SH4_SET_PMB 1,0x88,(BASE+0x08),128,0,0,1

All such macro instantiations have now been converted to use the
comma-delimited syntax, so that "BASE" can even be defined using
parenthesis.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>