[SH] Improved FLASH support on MB903 (STx7108M) board
This update improves support for FLASH on STMicroelectronics' MB903 (aka
STi7108M-HDK) board. This board has both NOR and SPI Serial flash on it
(there is no NAND flash).

The NOR flash is 128MiB, however, by default the targetpacks map the
entire 128MiB as CSA (EMI Bank #0). However, the largest possible EMI
bank is only 64MiB, hence only the first half of the NOR flash was usable.
The board actually allows both CSA and CSB to control the enable for the
NOR flash. This update defines both of the first two EMI Banks (#0 and #1)
to each be 64MiB, hence mapping 64MiB on each of CSA and CSB. This allows
the entire NOR flash to be readily accessed.

In addition, when in boot-from-SPI mode, prior to this patch *none*
of the NOR flash was usable, and hence U-Boot failed to "probe" the
NOR device correctly. Now that both EMI Banks #0 and #1 are correctly
configured to be 64MiB, is is possible in boot-from-SPI to see the upper
64MiB of the NOR flash.  It is not possible to see any of the lower 64MiB
of the NOR flash. If sharing code in the NOR flash between boot-from-NOR
and boot-from-SPI scenarios, please ensure you only use the upper half
of the NOR flash.

This was tested on a single board (MB903B-01 N067), with cut 2.0. Both
boot-from-NOR and boot-from-SPI environments were successfully tested
with all 128MiB of NOR being present in the former, but only the upper
64MiB of NOR present in the latter case.

In order to boot successfully, it is required to re-run the "romgen"
utility, and comment out the PMB setup, as per the instructions in
board/st/mb903/init-mb903.S.

Finally, it appears that in order to boot successfully (at least on the
board in question) from FLASH, that the JTAG cable must be unplugged!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>